src/requirements.txt
python-whois
waitress
flask
dnspython
blowfish